[One .NET] AOT support
Helpful reading:

* https://github.com/dotnet/runtime/blob/15dec9a2aa5a4236d6ba70de2e9c146867b9d2e0/src/tasks/AotCompilerTask/MonoAOTCompiler.cs
* https://github.com/dotnet/runtime/blob/15dec9a2aa5a4236d6ba70de2e9c146867b9d2e0/src/mono/netcore/nuget/Microsoft.NET.Runtime.MonoAOTCompiler.Task/README.md

TODO / limitations:

* We only have the osx-x64 host packages
* I have not updated the installers yet. The AOT packs are placed in
  `~/android-toolchain/dotnet` for local testing.
* Many other `TODO` comments

To make this work, I moved the existing `<Aot/>` MSBuild task calls to
a new `_AndroidAot` MSBuild target in `Xamarin.Android.Legacy.targets`.

In the .NET 6 targets, there is a *different* `_AndroidAot` MSBuild
target that runs the new `<MonoAOTCompiler/>` MSBuild task.

In order to acquire the AOT packages, I created a new
`mono-aot-compiler.proj` that is invoked such as:

    <Exec
        Command="$(DotNetPreviewTool) build @(_GlobalProperties, ' ') &quot;$(MSBuildThisFileDirectory)..\mono-aot-compiler\mono-aot-compiler.proj&quot;"
        EnvironmentVariables="NUGET_PACKAGES=$(DotNetPreviewPath)packs"
    />

Setting `$NUGET_PACKAGES` allows us to extract the AOT NuGet packages
directly to `dotnet/packs` when this project is restored.

Then we conditionally import the Sdk:

    <Import Project="..\build\Microsoft.NET.Runtime.MonoAOTCompiler.Task.props" Sdk="Microsoft.NET.Runtime.MonoAOTCompiler.Task"
        Condition=" '$(AotAssemblies)' == 'true' " />

Which allows us to use `$(MonoAOTCompilerTaskAssemblyPath)`:

    <UsingTask
        Condition=" '$(AotAssemblies)' == 'true' "
        TaskName="MonoAOTCompiler"
        AssemblyFile="$(MonoAOTCompilerTaskAssemblyPath)"
    />

I still need to add the new packs to our installers.